requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pytest==8.2.2
httpx==0.27.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
asyncpg==0.29.0